composer: use key rather than keycode for shortcut
The code property is used to represent the physical location of a key on
the keyboard, rather than the character created by pressing the key.

The mnenomic value of the shortcuts is lost if they follow the key
positions rather than characters - as well as causing conflicts with
existing shortcuts in moderately common keyboard layouts (cut and copy
on a dvorak layout).